Fix edge case in object.GetType dataflow
