+++ src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/Dataflow/HandleCallAction.cs
@@ -373,6 +373,11 @@ namespace ILLink.Shared.TrimAnalysis
                 //
                 case IntrinsicId.Object_GetType:
                     {
+                        if (instanceValue.IsEmpty ()) {
+                            AddReturnValue (MultiValueLattice.Top);
+                            break;
+                        }
+
                         foreach (var valueNode in instanceValue.AsEnumerable ())
                         {
                             // Note that valueNode can be statically typed in IL as some generic argument type.
